+++ src/services/irc.ts
@@ -76,54 +76,4 @@ export class IRCServer {
         });
     }
 
-    // Function that parses irc messages.
-    // Shamelessly stolen from node-irc https://github.com/aredridel/node-ircd
-    // Altered because I had to fight Typescript a bit :(
-    private parseMessage(line: string): IrcMessage {
-        const message: IrcMessage = {
-            params: [],
-            sender: "",
-            command: "",
-            error: undefined,
-        };
-        let m = /(:[^ ]+ )?([A-Z0-9]+)(?: (.*))?/i.exec(line);
-        if (!m) {
-            message["error"] = "Unable to parse message";
-            return message;
-        } else {
-            let i;
-            if (m[3] && (i = m[3].indexOf(':')) !== -1) {
-                let rest = m[3].slice(i + 1);
-                message.params = m[3].slice(0, i - 1).split(' ');
-                message.params.push(rest);
-            } else {
-                if (m[3]) {
-                    message.params = m[3].split(' ');
-                } else {
-                    message.params = [];
-                }
-            }
-            if (m[2]) {
-                message.command = m[2].toUpperCase();
-            }
-            if (m[1]) {
-                message.sender = m[1];
-            }
-
-            return message;
-        }
-    }
 }
-
-interface NetOptions {
-    allowHalfOpen?: boolean;
-    pauseOnConnect?: boolean;
-}
-
-interface IrcMessage {
-    command?: string;
-    params?: string[];
-    sender?: string;
-    error?: string;
-}
-
