+++ src/types.ts
@@ -1,0 +1,11 @@
+export interface IrcMessage {
+    command?: string;
+    params?: string[];
+    sender?: string;
+    error?: string;
+}
+
+export interface NetOptions {
+    allowHalfOpen?: boolean;
+    pauseOnConnect?: boolean;
+}
